+++ genmonlib/mymsgqueue.py
@@ -20,7 +20,7 @@ from genmonlib.mythread import MyThread
 # ------------ MyMsgQueue class -------------------------------------------------
 class MyMsgQueue(MySupport):
     # ------------ MyMsgQueue::init----------------------------------------------
-    def __init__(self, config=None, log=None, debug=False,callback=None):
+    def __init__(self, config=None, log=None, debug=False,callback=None, minimum_wait_between_messages = 0):
         super(MyMsgQueue, self).__init__()
         self.log = log
         self.config = config
@@ -31,6 +31,9 @@ class MyMsgQueue(MySupport):
 
         self.max_retry_time = 600  # 10 min
         self.default_wait = 120  # 2 min
+        self.minimum_wait_between_messages = minimum_wait_between_messages
+        self.last_message_sent_time = None
+
         self.debug = debug
 
         if self.config != None:
@@ -41,10 +44,13 @@ class MyMsgQueue(MySupport):
                 self.default_wait = self.config.ReadValue(
                     "default_wait", return_type=int, default=120
                 )
+                self.minimum_wait_between_messages = self.config.ReadValue(
+                    "minimum_wait_between_messages", return_type=int, default=minimum_wait_between_messages
+                )
                 self.debug = self.config.ReadValue(
                     "debug", return_type=bool, default=False
                 )
-
+                self.LogDebug("Min Wait Between Messages: " + str(self.minimum_wait_between_messages))
             except Exception as e1:
                 self.LogErrorLine(
                     "Error in MyMsgQueue:init, error reading config: " + str(e1)
@@ -65,6 +71,11 @@ class MyMsgQueue(MySupport):
             while self.MessageQueue != []:
                 messageError = False
                 try:
+                    message_time = datetime.datetime.now()
+                    if self.last_message_sent_time != None and ((message_time - self.last_message_sent_time).total_seconds() < self.minimum_wait_between_messages):
+                        if self.WaitForExit("QueueWorker", self.minimum_wait_between_messages):
+                            return
+                        continue
                     with self.QueueLock:
                         MessageItems = self.MessageQueue.pop()
                     if len(MessageItems[1]):
@@ -72,9 +83,7 @@ class MyMsgQueue(MySupport):
                     else:
                         ret_val = self.callback(MessageItems[0])
                     if not (ret_val):
-                        self.LogError(
-                            "Error sending message in QueueWorker, callback failed, retrying"
-                        )
+                        self.LogError("Error sending message in QueueWorker, callback failed, retrying")
                         messageError = True
                 except Exception as e1:
                     self.LogErrorLine("Error in QueueWorker, retrying (2): " + str(e1))
@@ -95,6 +104,10 @@ class MyMsgQueue(MySupport):
                                 return
                         else:
                             self.LogDebug("Message retry expired: " + MessageItems[0])
+                    else:
+                        # message was sent
+                        self.LogDebug("Message Sent")
+                        self.last_message_sent_time = message_time
                 except Exception as e1:
                     self.LogErrorLine(
                         "Error in QueueWorker requeue, retrying (3): " + str(e1)